fix(ffmpeg): align syntax of concatenation on the ffmpeg 2 and 3 version
In version 3.0.2, the concat operation doesn't working with FILE_1|FILE_2|FILE_3 syntax, so we have to create an intermediate file which contains all the information about the source to be concatenated. This modification is linked to the need of a better implementation for ffmpeg command line call.

Linked to #68 and waiting for https://github.com/bramp/ffmpeg-cli-wrapper/pull/22