feat(downloader.youtube): support video above 720p
Thanks to https://github.com/axet/vget lib, we can now fetch multiple stream from youtube, one for audio and one for video. After downloading each stream, we have to use https://github.com/bramp/ffmpeg-cli-wrapper/ lib to merge both stream into a unified video stream. The output format choosen is a video without encoding in MP4 container with an audio in encoded in AAC (because WebM | flac isn't supported in mp4 container). The format MP4 was the simplest way to do it for keeping the video in the same format, without a re-encoding.
 Special thank to @bramp & @axet for this, can't be done without theirs work on each one lib.

Close #68