fix(multiPartFile): relativize reader position in multipart case
Like @andre77 demonstrates, there is a problem when a multi-range access is done, because we are working on the same InputStream for each call. To fix that, I relativize each Range in order to skip the correct amount of byte, based on the previous Range. This is possible because range should be in order.

Close #74