feat(gulli): support gulli as a finder to create a new podcast
We use the replays page of gulli to find the information about the show to be registered.
The OpenGraph meta information are useful to get information about description. The title is fetch from breadcrumb and the cover from the show list behind the replay.
We store the direct url to the player to allow download.
In the player url, we can find the playlist and the element which should be played, so we can find the direct download url to use in the upper HTTPDownloader. We choose to use the mp4 direct link to simplify the work and don't require ffmpeg in this step of work.

Closes #75