feat(db): extract db from the app
As a transition to micro-service architecture, I choose to split the DB out of the application. We still can change the url to look after a local database, but the main goal will be to have an external db. We can easily load a db from docker with the command `docker run -it --rm -p 8999:81 -p 1521:1521 -v /tmp/h2-db:/opt/h2-data oscarfonts/h2` for local dev. This solution allow to have access to the web-ui and can be accessed from multiple client (which was not possible from the embedded solution)