feat(dev): integrate the ui-v1 inside the skaffold dev env
This allows development with skaffold for the ui and javascript part. It's based on build outside the container then copy it inside a new image and then reload the pod on the backend. Some improvements could be to do this inside the docker container, but I would like to keep system simple if we are not working on the UI part. Having the hot-reload and not using it will consume a lot of resources for nothing.