fix(update): provide more information during update phase
The mimeType and length should be provided by the update phase. This help the system to have information when providing the RSS even if the item are not downloaded. An implmentation for RSSUpdater is provided and some work should be done to do this (with default value or real value) from other updaters.

Link to #73