feat(kodi): support kodi http source
In order to provide a better connection between the
Podcast-Server and `kodi`, a dedicated endpoint
is available under `/kodi/`, compliant with the
kodi format expected by this logic (see [source](https://github.com/xbmc/xbmc/blob/master/xbmc/filesystem/HTTPDirectory.cpp#L174).

The system provides a better UI than standard file
servers, but manage some mandatory cases from the
kodi expected format (😅). This is done as an
experimentation and will be polished in following
commits, with more tests and better code. The
idea is to test it for real to confirm the
feature works better than kodi-to-s3-source or
direct file access of the minio system.

Related to #244