fix(yt-dlp): add possibility to provide advance format for youtube video
The goal is to have the possibility to provide something like `bv+ba[format_note~='(?i)original'] / bv+ba` to download only original audio. It was not possible before
due to a strict parameter parsing. From now on, `'` are not evaluated as a parameter
encapsulation character. Future refactor will bring back this feature.